requirements/common.txt
Django==1.7
django-pipeline==1.3.26
django-app-namespace-template-loader==0.2
Pillow==2.6.1
git+git://github.com/Fantomas42/django-blog-zinnia.git@feature/django-1.7
git+git://github.com/django-blog-zinnia/zinnia-theme-bootstrap.git@dc70fb96c8f31e83eaa157d151b21aa33a82438a
django-tinymce==1.5.3
git+git://github.com/bopbopmara/Django-tinymce-filebrowser.git
sorl-thumbnail==12.1c
